Initial implementation of etcd provider
Summary:
Initial implementation, allowing the creation of keys in Etcd.

This is relevant to service discovery (ref T29) and specifically enables
automation of the configuration necessary for D44 to work.

Test Plan:
Ran against a test CoreOS cluster. Created, updated, and deleted
keys via Terraform, and also verified that if the key is deleted outside of
Terraform, it notices and re-creates it.

Reviewers: kian, #developers

Reviewed By: kian, #developers

Subscribers: kian

Maniphest Tasks: T29

Differential Revision: https://phabricator.paperg.com/D45